[u/f] 1. move `arch/x86_64/*` -> `arch/x86_64/bits/`       2. provide `__syscall` impl, which takes definite arguments and          return the original value returned by kernel       3. wrapper : separate architectures by dividing sub-folders       4. wrapper : new wrapper - `w_sigaction`

+++ src/include/sys/syscall.h
@@ -7,7 +7,7 @@ __BEGIN_DECLS
 
 #include <bits/syscall.h>
 
-long syscall(int num, ...);
+long syscall(long num, ...);
 
 __END_DECLS
